requirements.txt
# Core dependencies
fastapi==0.95.1
uvicorn==0.22.0
pydantic==2.6.3
numpy==1.26.4
python-multipart==0.0.6
typing-extensions>=4.5.0  # Replacing typingx

# Geospatial dependencies (optional for production)
# rasterio==1.3.9
# pyproj==3.6.1
geopandas>=0.14.3
shapely>=2.0.1

# Optional dependencies for AI models
openai==1.35.0
# torch==2.2.0
# transformers==4.38.2
# ultralytics==8.1.2  # For YOLOv8

# Data infrastructure
redis==5.0.1
confluent-kafka==2.3.0

# Utilities
python-dotenv==1.0.1
tqdm==4.66.2
requests==2.31.0

# CORS
cors==1.0.1

# Data Processing Dependencies
sentinelsat>=1.1.1
rasterio>=1.3.8
numpy>=1.24.3
scipy>=1.10.1
laspy>=2.3.0
pyproj>=3.6.0

# Added from the code block
numpy>=1.21.0
rasterio>=1.2.10
scipy>=1.7.0
requests>=2.26.0
python-dotenv>=0.19.0
pystac>=1.0.0
pystac-client>=0.7.0

# Scientific Computing
scikit-learn>=1.0.2

# Image processing
scikit-image>=0.18.0
rasterio>=1.2.0
opencv-python>=4.5.0

# LIDAR processing
laspy>=2.0.0
dask==2023.5.0
distributed==2023.5.0

# Machine learning
scikit-learn>=0.24.0
tensorflow>=2.6.0
torch>=1.9.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# API and web
fastapi>=0.68.0
uvicorn>=0.15.0
aiohttp>=3.8.0
websockets>=10.0

# Database
redis>=4.0.0
kafka-python>=2.0.0

# Testing
pytest>=6.2.0
pytest-asyncio>=0.15.0

# Authentication and security
kubernetes==25.3.0
jwt==1.3.1
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1

# Monitoring and System Diagnostics
psutil==5.9.5
redis==5.0.1
kafka-python==2.0.2

# Database Dependencies
sqlalchemy==2.0.15
asyncpg==0.27.0
aiosqlite==0.19.0
alembic==1.10.4
psycopg2-binary==2.9.6 